smartverse-be/requirements.txt
transformers==4.36.2
fastapi
uvicorn
torch
python-pptx
python-multipart
scikit-learn
pillow
pytesseract